tests/wpt/web-platform-tests/bluetooth/server/getPrimaryServices/services-not-found.https.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script src="/bluetooth/resources/bluetooth-helpers.js"></script>
<script>
'use strict';
const test_desc = 'Request for services in a device with no services. Reject ' +
    'with NotFoundError.';
const expected = new DOMException('No Services found in device.',
    'NotFoundError');

bluetooth_test(() => getEmptyHealthThermometerDevice()
    .then(({device}) => assert_promise_rejects_with_message(
        device.gatt.getPrimaryServices(),
        expected)),
    test_desc);
</script>
